+++ src/KSPTextureLoader/Utils/AllocatorUtil.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Runtime.InteropServices;
+using System.Threading;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Jobs;
 using Unity.Profiling;
+using UnityEngine;
 
 namespace KSPTextureLoader.Utils;
 
@@ -16,7 +18,22 @@ internal static unsafe class AllocatorUtil
     static readonly ProfilerMarker AllocMarker = new("Marshal.AllocHGlobal");
     static readonly ProfilerMarker FreeMarker = new("Marshal.FreeHGlobal");
 
+    private const long MB = 1024 * 1024;
+
     internal const Allocator HGlobal = (Allocator)32;
+    private static long allocMem = 0;
+    internal static long AllocMem => Interlocked.Read(ref allocMem);
+    internal static bool IsAboveWatermark
+    {
+        get
+        {
+            var watermark = Config.Instance.MaxTextureLoadMemory * MB;
+            if (watermark == 0)
+                return false;
+
+            return (ulong)AllocMem > Config.Instance.MaxTextureLoadMemory * MB;
+        }
+    }
 
     internal static NativeArray<T> CreateNativeArrayHGlobal<T>(
         int length,
@@ -28,6 +45,7 @@ internal static unsafe class AllocatorUtil
         var ptr = (void*)Marshal.AllocHGlobal(sizeof(T) * length);
         if (ptr is null)
             throw new OutOfMemoryException("failed to allocate an array using AllocHGlobal");
+        Interlocked.Add(ref allocMem, length);
 
         var array = NativeArrayUnsafeUtility.ConvertExistingDataToNativeArray<T>(
             ptr,
@@ -59,6 +77,7 @@ internal static unsafe class AllocatorUtil
         {
             using var scope = FreeMarker.Auto();
             Marshal.FreeHGlobal((IntPtr)array.GetUnsafePtr());
+            Interlocked.Add(ref allocMem, -array.Length);
         }
         else
             array.Dispose();
@@ -72,4 +91,10 @@ internal static unsafe class AllocatorUtil
         new DisposeJob<T>(array).Schedule(dependsOn);
         array = default;
     }
+
+    internal static WaitUntil WaitUntilMemoryBelowWatermark() =>
+        WaitUntilMemoryBelowWatermark(Config.Instance.MaxTextureLoadMemory * MB);
+
+    internal static WaitUntil WaitUntilMemoryBelowWatermark(ulong watermark) =>
+        new(() => watermark <= 0 || AllocMem <= 0 || (ulong)AllocMem < watermark);
 }